2. Fold DependencyScannerImpl into DependencyScanningWorker.cpp and DependencyScannerTool.cpp.
